chore(examples): lazy loads next/headers in fetchPage of auth example
