+++ examples/draft-preview/next-app/app/[slug]/page.tsx
@@ -2,9 +2,10 @@ import { draftMode } from 'next/headers'
 import { notFound } from 'next/navigation'
 
 import { Page } from '../../payload-types'
+import { fetchPage } from '../_api/fetchPage'
+import { fetchPages } from '../_api/fetchPages'
 import { Gutter } from '../_components/Gutter'
 import RichText from '../_components/RichText'
-import { fetchPage, fetchPages } from '../cms'
 
 import classes from './index.module.scss'
