+++ examples/draft-preview/next-app/app/_api/fetchPages.ts
@@ -1,0 +1,11 @@
+import type { Page } from '../../payload-types'
+
+export const fetchPages = async (): Promise<Page[]> => {
+  const pageRes: {
+    docs: Page[]
+  } = await fetch(`${process.env.NEXT_PUBLIC_CMS_URL}/api/pages?depth=0&limit=100`).then(res =>
+    res.json(),
+  ) // eslint-disable-line function-paren-newline
+
+  return pageRes?.docs ?? []
+}
